Feat: get personal opening class for a student

+++ src/main/java/com/pet/moduleregister/application/moduleClass/dto/usecases/OpeningClass.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -15,7 +16,7 @@ public class OpeningClass {
     private Long moduleClassId;
     private String moduleClassCode;
     private String moduleName;
-    private int numberOfCredits;
+    private BigDecimal numberOfCredits;
     private Lecturer lecturer;
     private List<Schedule> schedules;
